rename domain folders
